add api

+++ proyecto/weed/models.py
@@ -5,8 +5,8 @@ from django.db import models
 class Weed(models.Model):
   #Atributes
   #geoLoc = models.PointField()
-  latitud = models.IntegerField(blank = True, null = False, default = 0)
-  longitud = models.IntegerField(blank = True, null = False, default = 0)
+  latitud = models.FloatField(blank = True, null = False, default = 0)
+  longitud = models.FloatField(blank = True, null = False, default = 0)
   fecha = models.DateField(auto_now_add = True)
   weed = models.BooleanField(blank = True, null = False, default = False)
   comentario = models.CharField(max_length = 50, null = False, default = '')
